backend/requirements.txt
PyPDF2==3.0.1
requests==2.31.0
pillow==10.0.1
fitz==0.0.1.dev2
pymupdf==1.23.0